refactor(lsp): keep lifecycle changes within source limits (#3963)
