+++ crates/vize_maestro/src/ide/definition/module_specifier_tests.rs
@@ -114,6 +114,42 @@ fn resolves_scoped_wildcard_declaration_subpath() {
 }
 
 #[test]
+fn component_definition_rejects_deleted_target_but_keeps_open_unsaved_target() {
+    let dir = tempdir().unwrap();
+    let component_path = dir.path().join("Deleted.vue");
+    let source_path = dir.path().join("Parent.vue");
+    let source = r#"<script setup lang="ts">
+import Deleted from './Deleted.vue'
+</script>
+<template><Deleted /></template>
+"#;
+    let uri = Url::from_file_path(source_path).unwrap();
+    let component_uri = Url::from_file_path(component_path).unwrap();
+    let state = ServerState::new();
+    state
+        .documents
+        .open(uri.clone(), source.to_string(), 1, "vue".to_string());
+    state.update_virtual_docs(&uri, source);
+    let offset = source.find("Deleted />").unwrap();
+    let context = IdeContext::new(&state, &uri, offset).unwrap();
+
+    assert!(DefinitionService::definition(&context).is_none());
+
+    state.documents.open(
+        component_uri.clone(),
+        "<template />\n".to_string(),
+        1,
+        "vue".to_string(),
+    );
+    let GotoDefinitionResponse::Scalar(location) =
+        DefinitionService::definition(&context).expect("open target definition")
+    else {
+        panic!("component definition must be a scalar location");
+    };
+    assert_eq!(location.uri, component_uri);
+}
+
+#[test]
 fn resolves_relative_declarations_and_rejects_package_escape() {
     let workspace = tempdir().unwrap();
     let source = workspace.path().join("src/App.vue");
